refactor: modularize and decouple services in Group creation and authentication logic
- Separated authentication logic into AuthService and AuthAPI, adhering to Single Responsibility Principle.
- Created GroupAPI and GroupService to handle all interactions with group-related backend operations.
- Refactored CreateGroupPage to utilize the new GroupService, removing direct API calls from the component.
- Ensured that all service responsibilities are clearly defined and aligned with the principles of SOLID and Clean Code.

+++ app/Repositories/GroupRepository.php
@@ -36,19 +36,28 @@ class GroupRepository extends CrudRepository
      */
     public function afterSave($resource, $attributes)
     {  
-        $resource->footballers()->sync(data_get($attributes, 'footballers', []));  
-        $resource->schedules()->delete();
-        
-        
-        $schedules = data_get($attributes, 'schedules', []);
-        foreach ($schedules as $schedule) {
-            $resource->schedules()->create([
-                'day' => $schedule['day'],
-                'start_time' => $schedule['startTime'],
-                'closing_time' => $schedule['endTime'],
-                'modality_id' => $schedule['modality_id'],
-            ]);
+        if(!empty(data_get($attributes, 'footballers', [])))
+        {
+            $resource->footballers()->sync(data_get($attributes, 'footballers', []));  
         }
+        
+        if(!empty(data_get($attributes, 'schedules', [])))
+        {
+            $resource->schedules()->delete();      
+        
+            $schedules = data_get($attributes, 'schedules', []);
+            
+            foreach ($schedules as $schedule) 
+            {
+                $resource->schedules()->create([
+                    'day' => $schedule['day'],
+                    'start_time' => $schedule['startTime'],
+                    'closing_time' => $schedule['endTime'],
+                    'modality_id' => $schedule['modality_id'],
+                ]);
+            }            
+        }
+        
         
         
         return $resource;
